refactoring: move logic of page 'sets' to a another file
